Refactor: Move things into directories and delete unused files. Python 3.7 compatibility.
