--- sampleCode.py
@@ -1,21 +1,0 @@
-import random
-
-import matplotlib.pyplot as plt
-import numpy as np
-import time
-
-fig, ax = plt.subplots()
-line, = ax.plot(np.random.randn(100))
-plt.show(block=False)
-
-tstart = time.time()
-num_plots = 0
-while time.time()-tstart < 5:
-    line.set_ydata(np.random.randn(100))
-
-    # ax.draw_artist(ax.patch)
-    # ax.draw_artist(line)
-    fig.canvas.draw_idle()
-    fig.canvas.flush_events()
-    num_plots += 1
-print(num_plots)
